move the EntityPollRatingFactory in its own file
