+++ backend/tournesol/tests/factories/ratings.py
@@ -2,13 +2,12 @@ import factory
 from factory import fuzzy
 
 from core.tests.factories.user import UserFactory
-from tournesol.models import EntityPollRating, Poll
+from tournesol.models import Poll
 from tournesol.models import ratings as ratings_models
 from tournesol.tests.factories.entity import VideoFactory
 
 
 class ContributorRatingFactory(factory.django.DjangoModelFactory):
-
     class Meta:
         model = ratings_models.ContributorRating
 
@@ -19,20 +18,9 @@ class ContributorRatingFactory(factory.django.DjangoModelFactory):
 
 
 class ContributorRatingCriteriaScoreFactory(factory.django.DjangoModelFactory):
-
     class Meta:
         model = ratings_models.ContributorRatingCriteriaScore
 
     contributor_rating = factory.SubFactory(ContributorRatingFactory)
     criteria = "better_habits"
     score = fuzzy.FuzzyDecimal(-10, 10)
-
-
-class EntityPollRatingFactory(factory.django.DjangoModelFactory):
-
-    class Meta:
-        model = EntityPollRating
-
-    poll = factory.LazyAttribute(lambda n: Poll.default_poll())
-    entity = factory.SubFactory(VideoFactory)
-    tournesol_score = fuzzy.FuzzyDecimal(-10, 10)
